requirements/base.txt
# Django

django==3.2.23

# Postgres
psycopg2-binary==2.8.6

# Configuration
django-environ==0.7.0

# Forms
django-braces==1.15.0
django-crispy-forms==1.13.0

# Models
django-model-utils==4.2.0

# Images
Pillow==8.4.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.47.0

# Unicode slugification
unicode-slugify==0.1.5
django-autoslug==1.9.8

# Time zones support
pytz>=2021.3

# Redis support
django-redis==5.2.0
redis>=4.1.0

# celery
celery==5.2.3

# asset compression
django-compressor==3.1

# stripe
PyJWT==2.3.0

# email
django-celery-email==3.0.0
mailchimp==2.0.10

django-loginas==0.3.10

# Webpack to play nice with react
django-webpack-loader==1.4.1

# Django Rest Framework
djangorestframework==3.13.1

# introducing graphql
markdown==3.3.6
django-filter==21.1
graphene-django==3.0.0b7
graphene==3.0

# blog
django-tinymce==3.4.0

# Your custom requirements go here
auditlog3==1.0.1
six==1.16.0
django-bootstrap-v5==1.0.8
crispy-bootstrap5==0.6

# stripe
dj-stripe==2.5.1
